Modernize packaging (`pyproject.toml`) and update for Sopel 8
Uses the newer plugin template now, including the `.plugin` submodule as
entry point instead of the root package.

Updated readme & NEWS formats, dropped Sopel 7.x support (and also
Python versions < 3.8). Housekept just a tiny bit, not comprehensively.
