+++ COPYING
@@ -5,7 +5,7 @@
       distribute this package, provided that:
           * copyright notices are retained unchanged,
           * any distribution of this package, whether modified or not,
-      includes this license text.
+            includes this license text.
    2. Permission is hereby also granted to distribute binary programs
       which depend on this package. If the binary program depends on a
       modified version of this package, you are encouraged to publicly
