+++ NEWS
@@ -1,12 +1,12 @@
-0.4.0
-=====
+## Changelog
+
+### 0.4.0
 
 Added:
 * `.docs` command alias, for when `.rtfm` feels harsh (#2)
 
 
-0.3.0
-=====
+### 0.3.0
 
 Added:
 * Proper help output for `.rtfm` command (#1)
@@ -16,8 +16,7 @@ Changed:
 * Simplify retrieving output URL
 
 
-0.2.0
-=====
+### 0.2.0
 
 Added:
 * Caching behavior: The configured Sphinx inventory is fetched at startup, and
@@ -25,8 +24,7 @@ Added:
 * Bot admins can force an update with the new `.rtfmupdate` command
 
 
-0.1.1
-=====
+### 0.1.1
 
 Changed:
 * Empty/missing query now outputs the docs' base URL instead of an error
@@ -35,7 +33,6 @@ Fixed:
 * `rtfm_base` was left in `bot.memory` after `shutdown()`
 
 
-0.1.0
-=====
+### 0.1.0
 
 Initial release.
