+++ pyproject.toml
@@ -1,0 +1,52 @@
+[build-system]
+requires = ["setuptools>=63.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+platforms = ["Linux x86, x86-64"]
+
+[tool.setuptools.packages.find]
+include = ["sopel_rtfm", "sopel_rtfm.*"]
+namespaces = false
+
+[tool.setuptools.dynamic]
+readme = { file=["README.md", "NEWS"], content-type="text/markdown" }
+
+[project]
+name = "sopel-rtfm"
+version = "0.5.0.dev0"
+description = "A plugin to suggest documentation links when someone asks a basic question."
+
+authors = [
+  { name="dgw", email="dgw@technobabbl.es" },
+]
+
+license = { text="EFL-2.0" }
+dynamic = ["readme"]
+
+classifiers = [
+  "Intended Audience :: Developers",
+  "Intended Audience :: System Administrators",
+  "License :: Eiffel Forum License (EFL)",
+  "License :: OSI Approved :: Eiffel Forum License",
+  "Topic :: Communications :: Chat :: Internet Relay Chat",
+]
+keywords = [
+  "sopel",
+  "plugin",
+  "bot",
+  "irc",
+]
+
+requires-python = ">=3.8, <4"
+dependencies = [
+  "sopel>=8.0",
+  "sphobjinv~=2.1",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/sopel-irc/sopel-rtfm"
+"Bug Tracker" = "https://github.com/sopel-irc/sopel-rtfm/issues"
+
+[project.entry-points."sopel.plugins"]
+"rtfm" = "sopel_rtfm.plugin"
