--- setup.cfg
@@ -1,26 +1,0 @@
-[metadata]
-name = sopel-rtfm
-version = 0.4.0
-description = A plugin to suggest documentation links when someone asks a basic question.
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/sopel-irc/sopel-rtfm
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=7.0,<8
-    sphobjinv~=2.1
-
-[options.entry_points]
-sopel.plugins =
-    rtfm = sopel_rtfm
